+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### Changelog
 
 #### 1.2.0-beta.6 (1/12/18)
+* Added: Plugin icon when updating via Dashboard > Updates.
 * Changed: Convert sticky header from JS to CSS-only.
 * Changed: Move mai_header_before and mai_header_after hooks outside of site-header.
 * Changed: Mobile menu toggle now uses psuedo-elements for less markup.